Capa Repository para Producto

+++ TP2_Grupo1/src/main/java/ar/edu/unju/fi/service/IProductoService.java
@@ -10,7 +10,7 @@ public interface IProductoService {
 	Producto getProducto();
 	public List<Producto> getListaProductos();
 
-	public Producto getBuscarProducto(int codigo);
+	public Producto getBuscarProducto(long codigo);
 	public void guardarProducto(Producto producto);
 	public void modificarProducto(Producto producto);
